Front-end usuario versión final
