+++ frontend/src/components/Admin/Sidebar/Sidebar.scss
@@ -1,12 +1,12 @@
 .sideBar {
     width: 14rem;
-    height: calc(100vh);
     background-color: #D3D3D3;
     color: #3A3A3C;
     box-shadow: 2px 0 5px rgba(0, 0, 0, 0.15);
     display: flex;
     flex-direction: column;
     padding: 1rem 0;
+    height: 100%; /* Ocupará toda la altura del contenedor padre */
 
     ul {
         list-style-type: none;
