+++ frontend/src/components/Admin/UsuarioCreateForm/index.js
@@ -1,0 +1,1 @@
+export * from "./UsuarioCreateForm";
