--- frontend/src/components/Admin/UsuarioForm/UsuarioForm.scss
@@ -1,38 +1,0 @@
-.add-user-form {
-    &__input {
-      background-color: #f7f2e7 !important;
-      border-color: #c64d34 !important;
-      color: #6e6159 !important;
-  
-      &:focus {
-        border-color: #e2b13c !important;
-        background-color: #f7f2e7 !important;
-      }
-    }
-  
-    &__button {
-      background-color: #88b04b !important;
-      color: white !important;
-      font-size: 1rem;
-  
-      &:hover {
-        background-color: #c64d34 !important;
-      }
-    }
-  
-    &__links {
-      display: flex;
-      justify-content: space-between;
-      margin-top: 1rem;
-  
-      a {
-        color: #c64d34;
-        font-size: 0.9rem;
-  
-        &:hover {
-          color: #e2b13c;
-          text-decoration: underline;
-        }
-      }
-    }
-  }
