--- frontend/src/components/Admin/UsuarioForm/index.js
@@ -1,1 +1,0 @@
-export * from "./UsuarioForm";
