+++ frontend/src/components/Admin/UsuarioModal/UsuarioModal.scss
@@ -1,26 +1,26 @@
 .modal-form-usuario {
   .ui.modal {
-    max-width: 80%; /* Ajusta el ancho máximo del modal */
-    height: 80vh; /* Ajusta la altura del modal */
+    max-width: 80%;
+    height: 80vh;
     display: flex;
     flex-direction: column;
     justify-content: center;
     align-items: center;
-    padding: 20px; /* Espaciado interno del modal */
+    padding: 20px;
 
     & > .header {
-      text-align: center; /* Centra el texto del header */
-      font-size: 1.5em; /* Tamaño de texto ajustado */
-      font-weight: bold; /* Texto en negrita */
-      color: #333; /* Color del texto */
-      width: 100%; /* Asegura que ocupe todo el ancho del modal */
-      margin: 0; /* Elimina márgenes */
-      padding: 10px 0; /* Ajusta el espacio interno */
+      text-align: center;
+      font-size: 1.5em;
+      font-weight: bold;
+      color: #333;
+      width: 100%;
+      margin: 0;
+      padding: 10px 0;
     }
   }
 
   .header-title {
-    text-align: center; /* Centra el texto específicamente para esta clase */
+    text-align: center;
   }
 
   .close-button {
@@ -40,18 +40,18 @@
 
   .contenedor-modal {
     display: flex;
-    flex-direction: column; /* Asegura que los elementos se apilen verticalmente */
+    flex-direction: column;
     justify-content: center;
-    align-items: stretch; /* Expande el formulario para ocupar todo el ancho */
-    width: 100%; /* Ocupar todo el ancho del modal */
-    height: 100%; /* Asegura ocupar todo el espacio disponible */
-    padding: 20px; /* Espaciado interno del contenedor */
+    align-items: stretch;
+    width: 100%;
+    height: 100%;
+    padding: 20px;
 
     form {
-      width: 100%; /* Hace que el formulario ocupe todo el ancho del contenedor */
-      max-width: none; /* Desactiva cualquier limitación de ancho máximo */
-      padding: 20px; /* Espaciado interno del formulario */
-      box-sizing: border-box; /* Incluye padding en el cálculo del ancho */
+      width: 100%;
+      max-width: none;
+      padding: 20px;
+      box-sizing: border-box;
     }
   }
 }
