--- frontend/src/components/Admin/UsuarioTable/index.js
@@ -1,1 +1,0 @@
-export * from "./UsuarioTable";
