+++ frontend/src/components/Admin/UsuariosTable/index.js
@@ -1,0 +1,1 @@
+export * from "./UsuariosTable";
