+++ frontend/src/components/Admin/index.js
@@ -1,6 +1,6 @@
 export * from "./Sidebar";
-export * from "./UsuarioForm";
+export * from "./UsuarioCreateForm";
 export * from "./UsuarioUpdateForm";
 export * from "./UsuarioDeleteModal";
-export * from "./UsuarioTable";
+export * from "./UsuariosTable";
 export * from "./UsuarioModal";
